refactor to remove registry and keep minimal abstract interface
