+++ bittensor/__init__.py
@@ -232,7 +232,7 @@ from .mock.keyfile_mock import MockKeyfile as MockKeyfile
 from .mock.subtensor_mock import MockSubtensor as MockSubtensor
 from .mock.wallet_mock import MockWallet as MockWallet
 
-from . import subnets
+from .subnets import SubnetsAPI as SubnetsAPI
 
 configs = [
     axon.config(),
