--- bittensor/subnets/__init__.py
@@ -1,1 +1,0 @@
-from .registry_api import APIRegistry, register_handler, SubnetsAPI
